Added access rights for category management on the model side of things

+++ src/main/java/model/handlers/CategoryHandler.java
@@ -6,10 +6,12 @@ import model.repository.DAO.CategoriesDB;
 import model.repository.entities.CategoryEntity;
 
 import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
 
 public class CategoryHandler {
+
+    // Authorized roles
+    public static final String[] accessRoles = {"Admin", "Worker"};
 
     // Response codes
     public static final int EXCEPTION                   = 0;
@@ -19,8 +21,18 @@ public class CategoryHandler {
     public static final int DELETE_FAILURE              = 5;
     public static final int UPDATE_OK                   = 6;
     public static final int UPDATE_FAILURE              = 7;
+    public static final int ACCESS_DENIED               = 8;
 
-    public static int newCategory(String name) {
+    // Singleton used for initialization
+    private static final CategoryHandler instance = new CategoryHandler();
+
+    private CategoryHandler() {
+        AccessList.addAccessRights(CategoryHandler.class, accessRoles);
+    }
+
+    public static int newCategory(String name, String... access) {
+        if (!AccessList.validateAccess(CategoryHandler.class, access))
+            return ACCESS_DENIED;
         CategoryEntity newCategory = new CategoryEntity(name);
         try {
             if (CategoriesDB.insert(newCategory)) {
@@ -32,7 +44,10 @@ public class CategoryHandler {
         }
     }
 
-    public static int updateCategoryName(int id, String name, Date sessionTimestamp) {
+    public static int updateCategoryName(int id, String name, Date sessionTimestamp, String... access) {
+        if (!AccessList.validateAccess(CategoryHandler.class, access))
+            return ACCESS_DENIED;
+
         CategoryEntity toUpdate = new CategoryEntity(name);
         toUpdate.id = id;
             try {
@@ -45,7 +60,9 @@ public class CategoryHandler {
         }
     }
 
-    public static int deleteCategory(int id) {
+    public static int deleteCategory(int id, String... access) {
+        if (!AccessList.validateAccess(CategoryHandler.class, access))
+            return ACCESS_DENIED;
         try {
             if (CategoriesDB.delete(id)) {
                 return DELETE_OK;
